// frontend/src/app/learncoreservices/models/platform-staff-details-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface PlatformStaffDetailsResponse {
  accountLocked?: boolean;
  avatarUrl?: string;
  createdDate?: string;
  email?: string;
  enabled?: boolean;
  firstname?: string;
  lastname?: string;
  permissions?: Array<string>;
  positionName?: string;
  roles?: Array<string>;
  uuid?: string;
}