// frontend/src/app/learncoreservices/models/staff-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface StaffResponse {
  accountLocked?: boolean;
  email?: string;
  enabled?: boolean;
  firstname?: string;
  lastModifiedDate?: string;
  lastname?: string;
  platformStaff?: boolean;
  positionName?: string;
  roles?: Array<string>;
  tenantUuid?: string;
  uuid?: string;
}